refactor: réorganisation de la structure des fichiers (avec maj des importations)
