--- composables/useDatabase.ts
@@ -1,5 +1,0 @@
-import { db } from '~/utils/database'
-
-export function useDatabase() {
-  return db
-}
