+++ server/api/[...auth].ts
@@ -1,4 +1,4 @@
-import { auth } from '~/utils/auth'
+import { auth } from '~/server/config/auth'
 
 export default defineEventHandler((event) => {
   return auth.handler(toWebRequest(event))
