+++ server/database/database.ts
@@ -1,5 +1,5 @@
 import { drizzle } from 'drizzle-orm/better-sqlite3'
-import * as schema from '../server/database/schema'
+import * as schema from './schema'
 
 const databaseUrl = 'database.sqlite'
