+++ server/utils/migration.ts
@@ -1,6 +1,6 @@
 import { consola } from 'consola'
 import { migrate } from 'drizzle-orm/better-sqlite3/migrator'
-import { db } from '../../utils/database'
+import { db } from '../database/database'
 
 export async function applyMigrationsIfNeeded() {
   consola.info('Apply database migrations if needed...')
