+++ server/utils/useDatabase.ts
@@ -1,0 +1,5 @@
+import { db } from '~/server/database/database'
+
+export function useDatabase() {
+  return db
+}
